requirements.txt
appdirs==1.4.3
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
cloudpickle==1.6.0
colorama==0.4.3
contextlib2==0.6.0
cycler==0.10.0
distlib==0.3.0
distro==1.4.0
future==0.18.2
gym==0.18.0
gym-retro==0.8.0
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
kiwisolver==1.3.1
lockfile==0.12.2
matplotlib==3.4.1
msgpack==0.6.2
neat-python==0.92
numpy==1.20.2
opencv-python==4.5.1.48
packaging==20.3
pep517==0.8.2
Pillow==7.2.0
progress==1.5
pyglet==1.4.5
pyparsing==2.4.7
python-dateutil==2.8.1
pytoml==0.1.21
requests==2.22.0
retrying==1.3.3
scipy==1.6.2
six==1.15.0
urllib3==1.25.8
webencodings==0.5.1
